+++ lib/src/bones_api_logging_io.dart
@@ -90,6 +90,12 @@ class LoggerHandlerIO extends LoggerHandler {
   }
 
   @override
+  bool forceFlushMessages() {
+    if (_printMessageQueue.isEmpty) return false;
+    return _flushPrintMessageQueue();
+  }
+
+  @override
   void printMessage(logging.Level level, String message) {
     if (!LoggerHandler.useLogQueue) {
       if (_printMessageQueueLevel != null) {